Reorganizes the project directory hierarchy to improve structure and maintainability
