+++ src/app/layout.js
@@ -1,5 +1,5 @@
 import { Geist, Geist_Mono } from "next/font/google";
-import "./globals.css";
+import "../styles/globals.css";
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
@@ -13,6 +13,9 @@ const geistMono = Geist_Mono({
 
 export const metadata = {
   title: "RPGo!",
+  icons: {
+    icon: "/img/icone.png",
+  },
   description: "Generated by create next app",
 };
