+++ src/app/signUp/page.js
@@ -1,31 +1,36 @@
-import Button from "../components/button";
-import InputField from "../components/inputField";
-import LogoWithTitle from "../components/logoWithTitle";
-import SignUpLink from "../components/signUpLink";
-import EmailIcon from "../lib/icons/EmailIcon";
-import PasswordIcon from "../lib/icons/PasswordIcon";
-import UserIcon from "../lib/icons/UserIcon";
+import Button from "../../components/button";
+import InputField from "../../components/inputField";
+import LogoWithTitle from "../../components/logoWithTitle";
+import SignUpLink from "../../components/signUpLink";
+import EmailIcon from "../../lib/icons/EmailIcon";
+import PasswordIcon from "../../lib/icons/PasswordIcon";
+import UserIcon from "../../lib/icons/UserIcon";
+import HeaderMenu from "@/components/hearderMenu";
 
 export default function SignUpPage() {
   return (
-    <div className="flex items-center justify-center h-screen bg-[#D4A59A]">
-      <div className="w-full max-w-sm p-6">
-        {/* Reutilizando o componente de logo e título */}
-        <LogoWithTitle />
+    <div className="bg-[#D4A59A] pt-5">
+      <HeaderMenu />
 
-        {/* Formulário de cadastro */}
-        <form className="space-y-6">
-          {/* Campos do formulário */}
-          <InputField type="text" placeholder="USERNAME" icon={<UserIcon />} />
-          <InputField type="email" placeholder="EMAIL" icon={<EmailIcon />} />
-          <InputField type="password" placeholder="PASSWORD" icon={<PasswordIcon />} />
-          <InputField type="confirmPassword" placeholder="CONFIRM PASSWORD" icon={<PasswordIcon />} />
-          {/* Botão Criar Conta */}
-          <Button label="CRIAR CONTA" variant="primary" />
-        </form>
+      <div className="flex items-center justify-center h-screen bg-[#D4A59A]">
+        <div className="w-full max-w-sm p-6">
+          {/* Reutilizando o componente de logo e título */}
+          <LogoWithTitle />
 
-        {/* Link para login */}
-        <SignUpLink />
+          {/* Formulário de cadastro */}
+          <form className="space-y-6">
+            {/* Campos do formulário */}
+            <InputField type="text" placeholder="USERNAME" icon={<UserIcon />} />
+            <InputField type="email" placeholder="EMAIL" icon={<EmailIcon />} />
+            <InputField type="password" placeholder="PASSWORD" icon={<PasswordIcon />} />
+            <InputField type="confirmPassword" placeholder="CONFIRM PASSWORD" icon={<PasswordIcon />} />
+            {/* Botão Criar Conta */}
+            <Button label="CRIAR CONTA" variant="primary" />
+          </form>
+
+          {/* Link para login */}
+          <SignUpLink />
+        </div>
       </div>
     </div>
   );
