+++ src/components/logoWithTitle.js
