+++ src/lib/icons/EmailIcon.js
