+++ src/lib/icons/PasswordIcon.js
