+++ src/lib/icons/UserIcon.js
